+++ frontend/src/utils/generate-warnings.ts
@@ -1,23 +1,16 @@
 import {
   Schedule,
-  ScheduleCourse,
-  INEUAndPrereq,
-  INEUOrPrereq,
-  ScheduleTerm,
-  IInitialScheduleRep,
-  INEUCourse,
-  ICompleteCourse,
-  ScheduleYear,
-  Season,
-  Status,
-  INEUParentMap,
-  CourseTakenTracker,
   IWarning,
+  CourseTakenTracker,
+  ScheduleTerm,
+  ScheduleCourse,
+  ICompleteCourse,
   IRequiredCourse,
+  INEUCourse,
   INEUPrereqCourse,
   IScheduleCourse,
-  IOldRequirement,
-  INEUClassMap,
+  INEUAndPrereq,
+  INEUOrPrereq,
 } from "../models/types";
 
 /**
@@ -263,248 +256,6 @@ function checkSemesterOverload(
     return [];
   }
 }
-
-/**
- * converts from old schedule to new schedule
- * @param sched the old schedule.
- */
-export function oldToNew(
-  old: IInitialScheduleRep,
-  parent: INEUParentMap
-): Schedule {
-  const byTermId: {
-    termIds: number[];
-    termMap: { [key: number]: ScheduleCourse[] };
-  } = {
-    termIds: [],
-    termMap: {},
-  };
-
-  // parse all courses in dude.
-  let allCourses: ICompleteCourse[] = [];
-  allCourses = allCourses.concat(old.completed.courses);
-  allCourses = allCourses.concat(old.inprogress.courses);
-  for (const course of allCourses) {
-    // if doesn't include termId, add.
-    if (!byTermId.termIds.includes(course.termId)) {
-      byTermId.termIds.push(course.termId);
-      byTermId.termMap[course.termId] = [];
-    }
-
-    // get the searchNEU version.
-    const detailed: INEUCourse | undefined = getSearchNEUData(course, parent);
-    if (detailed) {
-      byTermId.termMap[detailed.termId].push({
-        classId: detailed.classId,
-        subject: detailed.subject,
-        prereqs: detailed.prereqs,
-        coreqs: detailed.coreqs,
-        numCreditsMin: detailed.minCredits,
-        numCreditsMax: detailed.maxCredits,
-      });
-    } else {
-      // it probably exists, but might not in the termId.
-      const mostRecent: INEUCourse | undefined = getSearchNEUData(
-        {
-          subject: course.subject,
-          classId: course.classId,
-        },
-        parent
-      );
-
-      if (mostRecent) {
-        byTermId.termMap[course.termId].push({
-          classId: course.classId,
-          subject: course.subject,
-          prereqs: mostRecent.prereqs,
-          coreqs: mostRecent.coreqs,
-          numCreditsMin: mostRecent.minCredits,
-          numCreditsMax: mostRecent.maxCredits,
-        });
-      } else {
-        byTermId.termMap[course.termId].push({
-          classId: course.classId,
-          subject: course.subject,
-          prereqs: undefined,
-          coreqs: undefined,
-          numCreditsMin: course.creditHours,
-          numCreditsMax: course.creditHours,
-        });
-      }
-    }
-  }
-
-  // add to the born child.
-  const born: Schedule = {
-    years: [],
-    yearMap: {},
-    id: "example-schedule",
-  };
-
-  // get all the years
-  let years: number[] = [];
-  for (const termId of byTermId.termIds) {
-    if (!years.includes(Math.floor(termId / 100))) {
-      years.push(Math.floor(termId / 100));
-    }
-  }
-
-  // create each year object
-  for (const year of years) {
-    const yearObject: ScheduleYear = {
-      year: year,
-      fall: {
-        season: Season.FL,
-        year: year,
-        termId: year * 100 + 10,
-        id: 10,
-        status: Status.INACTIVE,
-        classes: [],
-      },
-      spring: {
-        season: Season.SP,
-        year: year,
-        termId: year * 100 + 30,
-        id: 30,
-        status: Status.INACTIVE,
-        classes: [],
-      },
-      summer1: {
-        season: Season.S1,
-        year: year,
-        termId: year * 100 + 40,
-        id: 40,
-        status: Status.INACTIVE,
-        classes: [],
-      },
-      summer2: {
-        season: Season.S2,
-        year: year,
-        termId: year * 100 + 60,
-        id: 60,
-        status: Status.INACTIVE,
-        classes: [],
-      },
-      isSummerFull: false,
-    };
-
-    if (byTermId.termMap[year * 100 + 10]) {
-      for (const course of byTermId.termMap[year * 100 + 10]) {
-        yearObject.fall.status = Status.CLASSES;
-        yearObject.fall.classes.push(course);
-      }
-    }
-    if (byTermId.termMap[year * 100 + 30]) {
-      for (const course of byTermId.termMap[year * 100 + 30]) {
-        yearObject.spring.status = Status.CLASSES;
-        yearObject.spring.classes.push(course);
-      }
-    }
-    if (byTermId.termMap[year * 100 + 40]) {
-      for (const course of byTermId.termMap[year * 100 + 40]) {
-        yearObject.summer1.status = Status.CLASSES;
-        yearObject.summer1.classes.push(course);
-      }
-    }
-    if (byTermId.termMap[year * 100 + 60]) {
-      for (const course of byTermId.termMap[year * 100 + 60]) {
-        yearObject.summer2.status = Status.CLASSES;
-        yearObject.summer2.classes.push(course);
-      }
-    }
-    born.years.push(year);
-    born.yearMap[year] = yearObject;
-  }
-
-  return born;
-}
-
-/**
- * Attempts to grab searchNEU data for a course, using that course's termId to lookup the corresponding file.
- * If no termId is found, automatically uses the most recent semester on record.
- * May return undefined.
- * @param course A course object (hopefully).
- * @param classMapParent The parent classMap object, with props "mostRecentSemester" and "allTermIds"
- * @returns Produces the corresponding searchNEU data for a class, if it exists. else => undefined.
- */
-export const getSearchNEUData = (
-  course:
-    | ICompleteCourse
-    | IRequiredCourse
-    | INEUCourse
-    | INEUPrereqCourse
-    | IScheduleCourse
-    | IOldRequirement,
-  classMapParent: INEUParentMap
-): INEUCourse | undefined => {
-  /**
-   * Grabs the data of a specified class.
-   * @param classObj The class object containing a classMap and termId.
-   * @param subject The subject (college abbreviation) of the target course.
-   * @param classId course number of the target course.
-   * @returns The resulting class object (if found).
-   */
-  function getClassData(
-    classObj: INEUClassMap,
-    potentialSubject: string,
-    potentialClassId: number
-  ) {
-    // classes can be accessed by the 'neu.edu/201830/<COLLEGE>/<COURSE_NUMBER>' attribute of each "classmap"
-    const query =
-      "neu.edu/" +
-      classObj.termId +
-      "/" +
-      potentialSubject +
-      "/" +
-      potentialClassId;
-    return classObj.classMap[query];
-  }
-
-  // skip doing work if there's no work to do.
-  if (!course) {
-    return undefined;
-  }
-
-  const subject: string | undefined = course.subject;
-  const classId: number | undefined = course.classId;
-  let termId: number | undefined;
-  let classMap: INEUClassMap | undefined;
-  if ("termId" in course) {
-    termId = course.termId;
-    classMap = classMapParent.classMapMap[termId];
-  }
-
-  if (classId && subject && termId && classMap) {
-    // if everything is valid, then query the classMap
-    // console.log("data found for: " + subject + classId);
-    return getClassData(classMap, subject, classId);
-  } else if (subject && classId && !termId) {
-    // if only the subject and classId are valid, guess the termId from most recent => least recent
-    const allTermIds = classMapParent.allTermIds;
-    for (const currentTermId of allTermIds) {
-      classMap = classMapParent.classMapMap[currentTermId];
-      const data: INEUCourse | undefined = getClassData(
-        classMap,
-        subject,
-        classId
-      );
-      if (data) {
-        // if the data exists, then return. otherwise keep searching.
-        // console.log("data found in term: " + termId + " for course: " + subject + classId);
-        return data;
-      }
-    }
-    // if not found, then return undefined
-    // console.log("data not found for:");
-    // console.log(course);
-    return undefined;
-  } else {
-    // if we have no subject and classId, then we don't even know what course to search for.
-    // console.log("data not found for:");
-    // console.log(course);
-    return undefined;
-  }
-};
 
 /**
  * Prereq Validation code follows
